// src/app/core/api/generated/model/periodicity.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const Periodicity = {
    Daily: 'daily',
    Weekly: 'weekly',
    Biweekly: 'biweekly',
    Monthly: 'monthly',
    Annual: 'annual'
} as const;
export type Periodicity = typeof Periodicity[keyof typeof Periodicity];