// src/app/core/api/generated/model/variableIncidenceType.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const VariableIncidenceType = {
    OnSalesPrice: 'on_sales_price',
    OnCost: 'on_cost',
    PerOrder: 'per_order',
    PerItem: 'per_item',
    PerProduct: 'per_product',
    PerChannel: 'per_channel'
} as const;
export type VariableIncidenceType = typeof VariableIncidenceType[keyof typeof VariableIncidenceType];